+++ TAMS.Entity/Concrete/TennisCourt.cs
@@ -1,14 +1,14 @@
 ﻿using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
 using TAMS.Common.Enums;
-using TAMS.Entity.Base;
+using TAMS.Entity.Abstracts;
 
 namespace TAMS.Entity.Concrete
 {
     public class TennisCourt : EntityBase
     {
         [Required]
-        public long ClubId { get; set; }
+        public int ClubId { get; set; }
 
         public string CourtName { get; set; }
